:recycle: cardコンポーネント修正
